% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{color_p_value}
\alias{color_p_value}
\title{Colorful P-Values}
\usage{
color_p_value(var_p_value)
}
\arguments{
\item{var_p_value}{Numeric variable of p-values to be colored conditionally.}
}
\value{
New character string variable with html code for color of p-value.
}
\description{
Color the p-values in html so they change colors conditional on values.
   Red indicates p-values less than 0.05.
Orange indicates p-values less than 0.15.
}
\examples{
color_p_value(c(0.00, 0.03, 0.05, 0.09, 0.11, 0.12, 0.16, 0.16, 0.21, 0.22))

}
