% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{plot_hist_continuous}
\alias{plot_hist_continuous}
\title{Histogram for Continuous Variables}
\usage{
plot_hist_continuous(df, var, n_binwidth, color_fill = "grey40")
}
\arguments{
\item{df}{Dataframe to be converted to table.}

\item{var}{Numeric for maximum number of digits to round numeric values.}

\item{n_binwidth}{Numeric indicating width of bins.}

\item{color_fill}{Default color for bars.}
}
\value{
Histogram.
}
\description{
Histogram for exploratory data analysis of a continuous variable.
}
\examples{
plot_hist_continuous(df = iris, var = Sepal.Length,  n_binwidth = 0.5, color_fill = "grey40")
plot_hist_continuous(df = USArrests, var = UrbanPop, n_binwidth = 0.5, color_fill = "red")

}
