% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{pretty_estimate}
\alias{pretty_estimate}
\title{Pretty Estimate}
\usage{
pretty_estimate(estimate, limit_low, limit_high, n_dec = 3)
}
\arguments{
\item{estimate}{Numeric estimate.}

\item{limit_low}{Numeric estimate lower confidence interval.}

\item{limit_high}{Numeric estimate higher confidence interval.}

\item{n_dec}{Numeric indicating number of digits to round after decimal.}
}
\value{
Character string of formatted estimate.
}
\description{
Create character string of an estimate with its confidence limits.
}
\examples{
iris_estimate   <- mean(iris$Petal.Length)
 iris_std_err   <- sd(iris$Petal.Length) / sqrt(length(iris$Petal.Length))
iris_limit_low  <- iris_estimate - 1.96 * iris_std_err
iris_limit_high <- iris_estimate + 1.96 * iris_std_err

pretty_estimate(iris_estimate, iris_limit_low, iris_limit_high, 2)
}
