% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{quick_kable}
\alias{quick_kable}
\title{Quick Kable}
\usage{
quick_kable(
  df,
  p_digits = 3,
  p_format = "html",
  lab_title = "",
  col_names = NULL
)
}
\arguments{
\item{df}{Dataframe to be converted to table.}

\item{p_digits}{Numeric for maximum number of digits to round numeric values.}

\item{p_format}{Character string. Possible values are latex, html, markdown, pandoc, and rst; this will be automatically determined if the function is called within knitr; it can also be set in the global option knitr.table.format. If format is a function, it must return a character string.}

\item{lab_title}{Character for table title.}

\item{col_names}{Character for column names. Must have the same length as the number of columns in the dataframe.}
}
\value{
The table ready for output in R Markdown.
}
\description{
Create table with default settings for kableExtra::kable() function.
Produces table in R Markdown.
}
\examples{
quick_kable(iris)
quick_kable(
    iris,
    lab_title = "Cool HTML Table of Iris Data",
    col_names = c("S Length", "S Width", "P Length", "P Width", "Species")
    )

}
