% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{set_theme_cowplot}
\alias{set_theme_cowplot}
\title{Set Cowplot Theme}
\usage{
set_theme_cowplot()
}
\value{
set theme
}
\description{
Set the cowplot theme, a minimal theme for ggplot2
}
\examples{
set_theme_cowplot()

}
