% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{std_var}
\alias{std_var}
\title{Standardize Variable}
\usage{
std_var(x)
}
\arguments{
\item{x}{Numeric variable to be formatted.}
}
\value{
The standardized \code{x}.
}
\description{
Standardize variable by centering on its mean and dividing by its standard deviation.
The new variable will now have a mean of 0 and a standard deviation of 1.
All missing data points are excluded from the mean and deviation calculations.
}
\examples{
std_var(iris$Petal.Length)
std_var(cars$speed)
}
