% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{trim_n}
\alias{trim_n}
\title{Trim Number}
\usage{
trim_n(x, n_dec = 1)
}
\arguments{
\item{x}{Numeric to be formatted.}

\item{n_dec}{Numeric for number of digits to format after the decimal.}
}
\value{
The formatted \code{x}.
}
\description{
Format number to n decimal places
}
\examples{
trim_n(pi, 3)
trim_n(1.2345, 1)

}
